+++ components/layout/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 
-import Navbar from '../navbar'
-import Footer from '../footer'
+import Navbar from '../shared/navbar'
+import Footer from '../shared/footer'
 
 export default function Layout({children, title = 'OpenTech', nav_foo = true}){
   return (
